requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]>=1.4
asyncpg
pydantic
python-dotenv
httpx
